FEATURE: Allow configuration of maximum varnish request header length
If a page with a huge number of tags is flushed, varnish might deny
the requests if the header is too large. By setting the headerLength
in the TagHandler, large requests will be split into multiple smaller
ones. 7500 is used as default value, as that is the default value
of the TagHandler if none is specified.